+++ java/integTest/src/test/java/glide/TransactionTestUtilities.java
@@ -72,6 +72,7 @@ public class TransactionTestUtilities {
 
         baseTransaction.unlink(new String[] {key3});
         baseTransaction.setrange(key3, 0, "GLIDE");
+        baseTransaction.getrange(key3, 0, 5);
 
         baseTransaction.hset(key4, Map.of(field1, value1, field2, value2));
         baseTransaction.hget(key4, field1);
@@ -106,6 +107,7 @@ public class TransactionTestUtilities {
         baseTransaction.smembers(key7);
         baseTransaction.sadd(setKey2, new String[] {"a", "b"});
         baseTransaction.sdiffstore(setKey3, new String[] {setKey2, key7});
+        baseTransaction.sinterstore(setKey3, new String[] {setKey2, key7});
 
         baseTransaction.zadd(key8, Map.of("one", 1.0, "two", 2.0, "three", 3.0));
         baseTransaction.zrank(key8, "one");
@@ -115,6 +117,7 @@ public class TransactionTestUtilities {
         baseTransaction.zmscore(key8, new String[] {"two", "three"});
         baseTransaction.zrange(key8, new RangeByIndex(0, 1));
         baseTransaction.zrangeWithScores(key8, new RangeByIndex(0, 1));
+        baseTransaction.zrangestore(key8, key8, new RangeByIndex(0, -1));
         baseTransaction.zscore(key8, "two");
         baseTransaction.zcount(key8, new ScoreBoundary(2, true), InfScoreBound.POSITIVE_INFINITY);
         baseTransaction.zlexcount(key8, new LexBoundary("a", true), InfLexBound.POSITIVE_INFINITY);
@@ -182,6 +185,7 @@ public class TransactionTestUtilities {
             0.5,
             1L,
             5L, // setrange(key3, 0, "GLIDE")
+            "GLIDE", // getrange(key3, 0, 5)
             2L,
             value1,
             2L, // hlen(key4)
@@ -209,8 +213,9 @@ public class TransactionTestUtilities {
             1L,
             true, // sismember(key7, "baz")
             Set.of("baz"),
-            2L, // sadd(setKey2, new String[] {"a", "b"})
+            2L, // sadd(setKey2, new String[] { "a", "b" })
             2L, // sdiffstore(setKey3, new String[] { setKey2, key7 })
+            0L, // sinterstore(setKey3, new String[] { setKey2, key7 })
             3L,
             0L, // zrank(key8, "one")
             4.0,
@@ -219,6 +224,7 @@ public class TransactionTestUtilities {
             new Double[] {2.0, 3.0}, // zmscore(key8, new String[] {"two", "three"})
             new String[] {"two", "three"}, // zrange
             Map.of("two", 2.0, "three", 3.0), // zrangeWithScores
+            2L, // zrangestore(key8, key8, new RangeByIndex(0, -1))
             2.0, // zscore(key8, "two")
             2L, // zcount(key8, new ScoreBoundary(2, true), InfScoreBound.POSITIVE_INFINITY)
             2L, // zlexcount(key8, new LexBoundary("a", true), InfLexBound.POSITIVE_INFINITY)